# docs/api/api.yaml
openapi: 3.0.3
info:
  title: spring-boot-jwt-template — Users API
  version: 0.1.0
  description: CRUD endpoints for user management and JWT authentication.

servers:
  - url: http://localhost:8080

paths:
  # ──────────────── Health ────────────────
  /syn:
    get:
      operationId: syn
      summary: Health check
      description: Returns syn-ack if the application is running.
      tags: [Health]
      responses:
        '200':
          description: Application is up
          content:
            text/plain:
              schema:
                type: string
              example: syn-ack

  # ──────────────── Authentication ────────────────
  /auth/register:
    post:
      operationId: register
      summary: Register a new user account (pending verification)
      description: Validates the registration data, saves the user with verified=false, generates a UUID verification token stored in Redis, and sends a verification link by email. The account is not active until verified via GET /auth/verification. The token expires after 15 minutes. Login is blocked until verification.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '202':
          description: Validation passed, verification link sent by email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBody'
              example:
                message: An email has been sent to verify your account
        '422':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 422
                error: Unprocessable Entity
                message: Password must be at least 8 characters
                timestamp: 2026-07-20T15:00:00Z
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 409
                error: Conflict
                message: Username is already taken
                timestamp: 2026-07-20T15:00:00Z
        '500':
          description: Failed to send verification email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 500
                error: Internal Server Error
                message: Failed to send verification email
                timestamp: 2026-07-20T15:00:00Z

  /auth/verification/{token}:
    get:
      operationId: verifyRegistration
      summary: Verify email and activate the account
      description: Validates the UUID verification token against the pending user record, sets verified=true, and returns a JWT token. The token expires after 15 minutes. This endpoint is also invoked after POST /auth/login and DELETE /users/{userId} (for account deletion confirmation).
      tags: [Auth]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID verification token sent by email
      responses:
        '200':
          description: Email verified, account activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyRegistrationResponse'
        '401':
          description: Invalid verification token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 401
                error: Unauthorized
                message: Invalid verification token
                timestamp: 2026-07-20T15:00:00Z

  /auth/login:
    post:
      operationId: login
      summary: Authenticate and request a verification link
      description: Validates credentials (username or email + password), generates a UUID verification token stored in Redis, and sends a verification link by email. The token expires after 15 minutes. Complete login via GET /auth/verification/{token}.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '202':
          description: Credentials validated, verification link sent by email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBody'
              example:
                message: A verification link has been sent to your email
        '422':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 422
                error: Unprocessable Entity
                message: Username or email is required
                timestamp: 2026-07-20T15:00:00Z
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 401
                error: Unauthorized
                message: Invalid credentials
                timestamp: 2026-07-20T15:00:00Z
        '403':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 403
                error: Forbidden
                message: Account has been locked
                timestamp: 2026-07-20T15:00:00Z

  /auth/resend-link:
    post:
      operationId: resendVerificationLink
      summary: Resend verification link
      description: Generates a new UUID verification token and sends a verification link by email. Use this when the previous token expired or was not received. The token expires after 15 minutes. Complete verification via GET /auth/verification/{token}.
      tags: [Auth]
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
            maxLength: 100
          description: Email address of the account to resend the verification link to
      responses:
        '202':
          description: Verification link regenerated and sent by email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBody'
              example:
                message: A verification link has been sent to your email
        '422':
          description: Email format validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 422
                error: Unprocessable Entity
                message: Email must be a valid email address
                timestamp: 2026-07-20T15:00:00Z
        '403':
          description: Email not found or account already verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 403
                error: Forbidden
                message: No pending verification found for this email
                timestamp: 2026-07-20T15:00:00Z
        '500':
          description: Failed to send verification email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 500
                error: Internal Server Error
                message: Failed to send verification email
                timestamp: 2026-07-20T15:00:00Z

  # ──────────────── GeoIP ────────────────
  /geoip:
    get:
      operationId: getOwnGeoIp
      summary: Resolve client geolocation
      description: Returns city, country, coordinates, timezone and other geolocation data for the calling client's IP. Powered by MaxMind GeoLite2-City.
      tags: [GeoIP]
      responses:
        '200':
          description: Geolocation data for the client's IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoIpResponse'
        '400':
          description: Invalid IP address format (e.g. from malformed X-Forwarded-For header)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 400
                error: Bad Request
                message: Invalid IP address format
                timestamp: 2026-07-25T12:00:00Z
        '404':
          description: IP address not found in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 404
                error: Not Found
                message: IP address not found in database
                timestamp: 2026-07-25T12:00:00Z

  /geoip/{ip}:
    get:
      operationId: getGeoIpByIp
      summary: Resolve IP geolocation
      description: Returns city, country, coordinates, timezone and other geolocation data for a given IP address. Powered by MaxMind GeoLite2-City.
      tags: [GeoIP]
      parameters:
        - name: ip
          in: path
          required: true
          schema:
            type: string
          description: IP address to look up (IPv4 or IPv6)
      responses:
        '200':
          description: Geolocation data for the resolved IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoIpResponse'
        '400':
          description: Invalid IP address format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 400
                error: Bad Request
                message: Invalid IP address format
                timestamp: 2026-07-25T12:00:00Z
        '404':
          description: IP address not found in database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 404
                error: Not Found
                message: IP address not found in database
                timestamp: 2026-07-25T12:00:00Z

  # ──────────────── Users CRUD ────────────────
  /users:
    get:
      operationId: listUsers
      summary: List users (paginated, admin only)
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
          description: Filter by username (partial match)
        - name: firstName
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
          description: Filter by first name (partial match)
        - name: lastName
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
          description: Filter by last name (partial match)
        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
            maxLength: 100
          description: Filter by email (partial match)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number (1-indexed)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Page size
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 401
                error: Unauthorized
                message: Missing or invalid authentication token
                timestamp: 2026-07-20T15:00:00Z
        '403':
          description: Insufficient permissions (admin role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 403
                error: Forbidden
                message: Admin role is required to access this resource
                timestamp: 2026-07-20T15:00:00Z
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 422
                error: Unprocessable Entity
                message: Page size must be between 1 and 100
                timestamp: 2026-07-20T15:00:00Z

  /users/{userId}:
    get:
      operationId: getUser
      summary: Get a user by ID
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 401
                error: Unauthorized
                message: Missing or invalid authentication token
                timestamp: 2026-07-20T15:00:00Z
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 403
                error: Forbidden
                message: Insufficient privileges to access this resource
                timestamp: 2026-07-20T15:00:00Z
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 404
                error: Not Found
                message: User not found
                timestamp: 2026-07-20T15:00:00Z

    put:
      operationId: updateUser
      summary: Update a user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 422
                error: Unprocessable Entity
                message: Email must be a valid email address
                timestamp: 2026-07-20T15:00:00Z
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 404
                error: Not Found
                message: User not found
                timestamp: 2026-07-20T15:00:00Z
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 409
                error: Conflict
                message: Email is already taken
                timestamp: 2026-07-20T15:00:00Z
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 422
                error: Unprocessable Entity
                message: Cannot change role of a SUPER_ADMIN user
                timestamp: 2026-07-20T15:00:00Z

    delete:
      operationId: deleteUser
      summary: Delete a user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                status: 404
                error: Not Found
                message: User not found
                timestamp: 2026-07-20T15:00:00Z

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRole:
      type: string
      enum: [admin, customer]

    # ── Auth ──
    RegisterInput:
      type: object
      required: [username, password, confirmPassword, firstName, lastName, email]
      properties:
        username:
          type: string
          maxLength: 50
        password:
          type: string
          minLength: 12
          maxLength: 255
          writeOnly: true
        confirmPassword:
          type: string
          writeOnly: true
          description: Must match password
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 100
      example:
        username: jdoe
        password: s3cur3P@ss
        confirmPassword: s3cur3P@ss
        firstName: John
        lastName: Doe
        email: jdoe@example.com

    LoginInput:
      type: object
      required: [password]
      properties:
        username:
          type: string
          description: Username to log in with
        email:
          type: string
          format: email
          maxLength: 100
          description: Email to log in with
        password:
          type: string
          writeOnly: true
      description: Provide either username or email (at least one required)
      example:
        username: jdoe
        password: s3cur3P@ss

    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT bearer token
        user:
          $ref: '#/components/schemas/User'
      example:
        token: eyJhbG...AwfQ...
        user:
          id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
          username: jdoe
          firstName: John
          lastName: Doe
          email: jdoe@example.com
          role: customer
          createdAt: 2026-07-20T14:00:00Z
          updatedAt: null

    VerifyRegistrationResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT bearer token
        user:
          $ref: '#/components/schemas/User'
      example:
        token: eyJhbG...AwfQ...
        user:
          id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
          username: jdoe
          firstName: John
          lastName: Doe
          email: jdoe@example.com
          role: customer
          createdAt: 2026-07-20T14:00:00Z
          updatedAt: null

    # ── Read model ──
    User:
      type: object
      required: [id, username, firstName, lastName, email, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          maxLength: 50
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 100
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      example:
        id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        username: jdoe
        firstName: John
        lastName: Doe
        email: jdoe@example.com
        role: customer
        createdAt: 2026-07-20T14:00:00Z
        updatedAt: null

    # ── Update input ──
    UpdateUserInput:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
        password:
          type: string
          minLength: 8
          maxLength: 255
          writeOnly: true
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 100
        role:
          $ref: '#/components/schemas/UserRole'
      example:
        firstName: Jonathan
        lastName: Doe

    # ── Simple message body ──
    MessageBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable message
      example:
        message: A verification link has been sent to your email

    # ── Shared error body ──
    ErrorBody:
      type: object
      required: [status, error, message, timestamp]
      properties:
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Short error type / code
        message:
          type: string
          description: Human-readable error description
        timestamp:
          type: string
          format: date-time
          description: When the error occurred

    # ── GeoIP ──
    GeoIpResponse:
      type: object
      required: [ip]
      properties:
        ip:
          type: string
          description: Resolved IP address
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        countryCode:
          type: string
          nullable: true
        continent:
          type: string
          nullable: true
        subdivision:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
      example:
        ip: 8.8.8.8
        city: Mountain View
        country: United States
        countryCode: US
        continent: North America
        subdivision: California
        postalCode: 94043
        timezone: America/Los_Angeles
        latitude: 37.386
        longitude: -122.084

    Meta:
      type: object
      required: [page, size, total]
      properties:
        page:
          type: integer
          description: Current page number (1-indexed)
        size:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items across all pages
      example:
        page: 1
        size: 20
        total: 1